requirements.txt
Flask
Flask-SQLAlchemy
gunicorn
werkzeug
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform
google-cloud-storage
python-dotenv
